+++ src/main/java/com/example/bank_app/service/UserService.java
@@ -5,7 +5,7 @@ import com.example.bank_app.dto.user.UserResponse;
 import com.example.bank_app.dto.user.UserUpdateRequest;
 
 public interface UserService {
-    UserResponse updateUser(UserUpdateRequest request, Integer userId);
-    void changePassword(ChangePasswordRequest request, Integer userId);
-    UserResponse getUserProfile(Integer userId);
+    UserResponse updateUser(UserUpdateRequest request, Long userId);
+    void changePassword(ChangePasswordRequest request, Long userId);
+    UserResponse getUserProfile(Long userId);
 }
